slight refactor of the TcpTransport. Extracted some useful common methods in preparation for adding 0mq.
